ajout absence côté user

+++ models/absence.manager.php
@@ -9,12 +9,12 @@ function getAllAbsence()
     return $stmt->fetchAll();
 }
 
-function getAllAbsenceByUserId($user_id)
+function getAllAbsenceByUserId()
 {
     require '../config/connect.php';
     $sql = 'SELECT * FROM table_absence WHERE user_id = :user_id';
     $stmt = $db->prepare($sql);
-    $stmt->bindValue(':user_id', $user_id);
+    $stmt->bindValue(':user_id', $_SESSION['user_id']);
     $stmt->execute();
     return $stmt->fetchAll();
 }
@@ -23,9 +23,14 @@ function addAbsence()
 {
     require '../config/connect.php';
     $sql = "INSERT INTO table_absence 
-    (absence_date_start, absence_date_end, absence_date_declaration,absence_type_id) 
-    VALUES (:absence_date_start, :absence_date_end, :absence_date_declaration,:absence_type_id) WHERE user_id = :user_id";
+    (absence_date_start, absence_date_end, absence_date_declaration, absence_type_id, user_id) 
+    VALUES (:absence_date_start, :absence_date_end, :absence_date_declaration, :absence_type_id, :user_id)";
     $stmt = $db->prepare($sql);
+    $stmt->bindValue(':absence_date_declaration', $_POST['absence_date_declaration']);
+    $stmt->bindValue(':user_id', $_SESSION['user_id']);
+    $stmt->bindValue(':absence_date_end', $_POST['absence_date_end']);
+    $stmt->bindValue(':absence_date_start', $_POST['absence_date_start']);
+    $stmt->bindValue(':absence_type_id', $_POST['absence_type_id']);
     $stmt->execute();
 }
 
@@ -38,21 +43,23 @@ function updateAbsence()
     absence_date_end = :absence_date_end,
     absence_date_declaration = :absence_date_declaration,
     absence_type_id = :absence_type_id
-    WHERE absence_id = :absence_id";
+    WHERE absence_id = :absence_id, user_id = :user_id";
 
     $stmt = $db->prepare($sql);
     $stmt->bindValue(':absence_date_start', $_POST['absence_date_start']);
     $stmt->bindValue(':absence_date_end', $_POST['absence_date_end']);
     $stmt->bindValue(':absence_date_declaration', $_POST['absence_date_declaration']);
     $stmt->bindValue(':absence_type_id', $_POST['absence_type_id']);
+    $stmt->bindValue(':user_id', $_SESSION['user_id']);
     $stmt->execute();
 }
 
 function deleteAbsence()
 {
     require '../config/connect.php';
-    $sql = "DELETE FROM table_absence WHERE absence_id = :absence_id";
+    $sql = "DELETE FROM table_absence WHERE absence_id = :absence_id, user_id = :user_id";
     $stmt = $db->prepare($sql);
     $stmt->bindValue(':absence_id', $_POST['absence_id']);
+    $stmt->bindValue(':user_id', $_SESSION['user_id']);
     $stmt->execute();
 }
